# server/src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Learning app",
        "description": "Free & modern app to learn and widen your horizonts, as well as social circles. All of this with surprisingly powerful flashcards",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [],
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "jwt"
            },
            "oAuth2AuthCode": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "https://accounts.google.com/o/auth2/v2/auth",
                        "tokenUrl": "https://accounts.google.com/o/auth2/v4/token",
                        "scopes": {
                            "profile:read": "Access to your profile",
                            "email:read": "Access to your email"
                        }
                    }
                }
            }
        },
        "schemas": {
            "Flashcard": {
                "required": ["isQuizQuestion", "prompt", "answers"],
                "type": "object",
                "properties": {
                    "prompt": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "extraInfo": {
                        "type": "string"
                    },
                    "correctAnswer": {
                        "type": "number",
                        "default": 0
                    },
                    "isQuizQuestion": {
                        "type": "boolean",
                        "default": false
                    }
                }
            },
...
            "FlashcardCollection": {
                "required": ["name", "owner"],
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "mongoose.Schema.Types.ObjectId"
                    },
                    "owner": {
                        "type": "string",
                        "description": "Ref to user's ID"
                    },
                    "name": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "flashcards": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "description": "Ref to flashcard's ID or flashcards' IDs"
                        }
                    },
                    "isPublic": {
                        "type": "boolean",
                        "default": false
                    }
                }
            },
            "NewFlashcardCollection": {
                "required": ["name"],
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "isPublic": {
                        "type": "boolean",
                        "default": false
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/api/flashcard/{flashcardId}": {
            "get": {
                "summary": "Display a flashcard",
                "tags": ["flashcard"],
                "description": "Find a single flashcard by ID and display its content",
                "operationId": "getFlashcardById",
                "parameters": [
                    {
                        "name": "flashcardId",
                        "in": "path",
                        "required": true,
                        "description": "Please provide the ID of the flashcard you wish to display",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Flashcard",
                                    "_id": {
                                        "type": "string",
                                        "description": "mongoose.Schema.Types.ObjectId"
                                    },
                                    "collectionId": {
                                        "type": "string",
                                        "description": "Ref to collection's ID"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID"
                    },
                    "401": {
                        "description": "User not logged in"
                    },
                    "403": {
                        "description": "You do not have permission to access this resource"
                    },
                    "404": {
                        "description": "The flashcard with the given ID was not found"
                    },
                    "500": {
                        "description": "Ooops! Something went wrong."
                    }
                }
            },
            "delete": {
                "summary": "Delete a flashcard",
                "tags": ["flashcard"],
                "parameters": [
                    {
                        "name": "flashcardId",
                        "in": "path",
                        "required": true,
                        "description": "Please provide the ID of the flashcard you wish to delete",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "operationId": "getFlashcardById",
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Invalid ID or bad request"
                    },
                    "401": {
                        "description": "User not logged in"
                    },
                    "403": {
                        "description": "You do not have permission to access this resource"
                    },
                    "404": {
                        "description": "The flashcard with the given ID was not found"
                    },

                    "500": {
                        "description": "Ooops! Something went wrong."
                    }
                }
            },
            "patch": {
                "summary": "Modify a flashcard",
                "tags": ["flashcard"],
                "parameters": [
                    {
                        "name": "flashcardId",
                        "in": "path",
                        "required": true,
                        "description": "Please provide the ID of the flashcard you wish to modify",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "operationId": "getFlashcardById",
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "requestBody": {
                    "description": "Modify a single flashcard in the given flashcard collection",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Flashcard"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Flashcard",
                                    "_id": {
                                        "type": "string",
                                        "description": "Ref to flashcard's ID"
                                    },
                                    "collectionId": {
                                        "type": "string",
                                        "description": "Ref to collection's ID"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID or bad request"
                    },
                    "401": {
                        "description": "User not logged in"
                    },
                    "403": {
                        "description": "You do not have permission to access this resource"
                    },
                    "404": {
                        "description": "The flashcard with the given ID was not found"
                    },

                    "500": {
                        "description": "Ooops! Something went wrong."
                    }
                }
            }
        },
        "/api/flashcard/{collectionId}": {
            "put": {
                "summary": "Create a flashcard in the given flashcard collection",
                "tags": ["flashcard"],
                "parameters": [
                    {
                        "name": "collectionId",
                        "in": "path",
                        "required": true,
                        "description": "Please provide the ID of the flashcard collection",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "operationId": "getCollectionById",
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "requestBody": {
                    "description": "Create a single flashcard in the given flashcard collection",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Flashcard"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Flashcard",
                                    "_id": {
                                        "type": "string",
                                        "description": "mongoose.Schema.Types.ObjectId"
                                    },
                                    "collectionId": {
                                        "type": "string",
                                        "description": "Ref to collection's ID"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID or bad request"
                    },
                    "401": {
                        "description": "User not logged in"
                    },
                    "403": {
                        "description": "You do not have permission to access this resource"
                    },
                    "404": {
                        "description": "The flashcard with the given ID was not found"
                    },

                    "500": {
                        "description": "Ooops! Something went wrong."
                    }
                }
            }
        },
        "/api/flashcard-collection/": {
            "post": {
                "summary": "Create a flashcard collection",
                "tags": ["flashcard collection"],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "requestBody": {
                    "description": "Create a single flashcard collection",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewFlashcardCollection"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlashcardCollection"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID or bad request"
                    },
                    "401": {
                        "description": "User not logged in"
                    },
                    "404": {
                        "description": "The flashcard collection with the given ID was not found"
                    },
                    "500": {
                        "description": "Ooops! Something went wrong."
                    }
                }
            }
        },
        "/api/flashcard-collection/{collectionId}": {
            "get": {
                "summary": "Display a flashcard collection",
                "tags": ["flashcard collection"],
                "description": "Display the content of a chosen flashcard collection",
                "operationId": "getCollectiondById",
                "parameters": [
                    {
                        "name": "collectionId",
                        "in": "path",
                        "required": true,
                        "description": "Please provide the ID of the flashcard collection you wish to display",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlashcardCollection"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID or bad request"
                    },
                    "401": {
                        "description": "User not logged in"
                    },
                    "403": {
                        "description": "You do not have permission to access this resource"
                    },
                    "404": {
                        "description": "The flashcard collection with the given ID was not found"
                    },
                    "500": {
                        "description": "Ooops! Something went wrong."
                    }
                }
            },
            "delete": {
                "summary": "Delete a flashcard collection",
                "tags": ["flashcard collection"],
                "operationId": "getCollectiondById",
                "parameters": [
                    {
                        "name": "collectionId",
                        "in": "path",
                        "required": true,
                        "description": "Please provide the ID of the flashcard collection you wish to delete",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Invalid ID or bad request"
                    },
                    "401": {
                        "description": "User not logged in"
                    },
                    "403": {
                        "description": "You do not have permission to access this resource"
                    },
                    "404": {
                        "description": "The flashcard collection with the given ID was not found"
                    },

                    "500": {
                        "description": "Ooops! Something went wrong.",
                        "content": {}
                    }
                }
            },
            "put": {
                "summary": "Modify a flashcard collection",
                "tags": ["flashcard collection"],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "requestBody": {
                    "description": "Modify a single flashcard collection",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewFlashcardCollection"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlashcardCollection"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID or bad request"
                    },
                    "401": {
                        "description": "User not logged in"
                    },
                    "404": {
                        "description": "The flashcard collection with the given ID was not found"
                    },
                    "500": {
                        "description": "Ooops! Something went wrong."
                    }
                }
            }
        },
...
    }
}